medintel-backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
pytesseract==0.3.10
Pillow==9.5.0
python-dotenv==1.0.0
scikit-learn==1.2.2
numpy==1.24.3
Werkzeug==2.2.3
SQLAlchemy==2.0.9
pytest==7.3.1
gunicorn==20.1.0